+++ core-java/src/test/java/com/baeldung/concurrent/countdownlatch/CountdownLatchExampleIntegrationTest.java
@@ -12,7 +12,7 @@ import java.util.stream.Stream;
 import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class CountdownLatchExampleTest {
+public class CountdownLatchExampleIntegrationTest {
     @Test
     public void whenParallelProcessing_thenMainThreadWillBlockUntilCompletion() throws InterruptedException {
         // Given
