+++ xmlunit2/pom.xml
@@ -17,6 +17,16 @@
                       <target>1.8</target>
                   </configuration>
               </plugin>
+              <plugin>
+                  <groupId>org.apache.maven.plugins</groupId>
+                  <artifactId>maven-surefire-plugin</artifactId>
+                  <configuration>
+                      <excludes>
+                          <exclude>**/*IntegrationTest.java</exclude>
+                          <exclude>**/*LiveTest.java</exclude>
+                      </excludes>
+                  </configuration>
+              </plugin>
           </plugins>
       </build>
 
@@ -55,6 +65,6 @@
 
         <!-- Maven plugins -->
         <maven-compiler-plugin.version>3.6.0</maven-compiler-plugin.version>
-
     </properties>
+
 </project>
